+++ package.json
@@ -38,8 +38,8 @@
   "workspaces": {
     "catalogs": {
       "db": {
-        "drizzle-orm": "1.0.0-beta.19",
-        "drizzle-kit": "^0.31.10"
+        "drizzle-orm": "1.0.0-beta.20",
+        "drizzle-kit": "1.0.0-beta.20"
       },
       "discord": {
         "@sapphire/discord.js-utilities": "^7.3.3",
